# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PatentFlow BE API",
    "description" : "PatentFlow FE 우선 연동 API 문서",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "paths" : {
...
    "/api/v1/legal/dashboard/area-distribution" : {
      "get" : {
        "tags" : [
          "legal-controller"
        ],
        "operationId" : "getAreaDistribution",
        "parameters" : [
          {
            "name" : "quarter",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "country",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "dateFrom",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "date"
            }
          },
          {
            "name" : "dateTo",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "date"
            }
          },
          {
            "name" : "reviewWorkflowStatus",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "enum" : [
                "NOT_IN_REVIEW",
                "REVIEW_QUARTER_STARTED",
                "MAIL_READY",
                "WAITING_BUSINESS_RESPONSE",
                "BUSINESS_RESPONSE_RECEIVED"
              ]
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAreaDistributionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/legal/audit-logs" : {
      "get" : {
        "tags" : [
          "legal-controller"
        ],
        "operationId" : "getAuditLogs",
        "parameters" : [
          {
            "name" : "type",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "patentId",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "limit",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 100
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAuditLogEntryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseAreaDistributionResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/AreaDistributionResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AreaDistributionResponse" : {
        "type" : "object",
        "properties" : {
          "totalCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "businessArea" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AreaGroupResponse"
            }
          },
          "technologyArea" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AreaGroupResponse"
            }
          },
          "product" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AreaGroupResponse"
            }
          }
        }
      },
      "AreaGroupResponse" : {
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          },
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "relatedLabels" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ApiResponseListAuditLogEntryResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditLogEntryResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AuditLogEntryResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "patentId" : {
            "type" : "string"
          },
          "actor" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}